+++ changelog.md
@@ -3,6 +3,7 @@ vX.X.X
 
 - Disabled metric prefixes for IPC-2152 calculator variables, closes #118.
 - Fixed info section of "Via Current" calculator, closes #112.
+- Added ability to display closest higher value and closest lower value on the standard resistor finder, closes #120.
 
 v1.2.0
 ------